pyproject.toml
[project]
name = "banshee-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "py-cord>=2.6.1",
    "python-dotenv>=1.2.1",
    "tortoise-orm>=0.25.1",
]
